backend/requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Pipecat — voice pipeline
pipecat-ai[google,deepgram,silero,daily]>=1.0.0
daily-python>=0.27.0
aiohttp>=3.9.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0

# Auth
pyjwt>=2.12.0
resend==2.28.0

# LLM providers (for compiler and profiler — runtime uses Pipecat's GoogleLLMService)
anthropic>=0.42.0

# Storage
boto3>=1.35.0

# File parsing (LO extraction from uploaded syllabi/documents)
PyMuPDF>=1.26.0
python-docx>=1.0.0

# Audio filtering (echo/noise suppression for TTS)
pyrnnoise>=0.4.0

# Utilities
loguru>=0.7.0
pydantic>=2.0.0
